+++ constant/log-prefix.js
@@ -1,0 +1,14 @@
+/**
+ * @since 2016-08-21 10:23
+ * @author vivaxy
+ */
+
+'use strict';
+
+/* eslint-disable */
+exports.BROWSER  = ' browser:';
+exports.SERVER   = '  server:';
+exports.WATCH    = '   watch:';
+exports.RESPONSE = 'response:';
+exports.REQUEST  = ' request:';
+exports.TIME     = '    time:';