+++ CHANGELOG.md
@@ -1,6 +1,13 @@
-# v3.1.1
+# v3.2.0
 
+- support restful API, config same as `koa-router`
 - upgrade jade to pug
+- update node versions in travis ci
+- add verbose log for request method, path and cost time
+- fix directory config
+- change server response log from debug level to verbose
+- add ssl test case
+- update readme
 
 # v3.1.0